Extract Home headers to useHomeHeaders hook
